requirements.txt
asgiref==3.2.10
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
defusedxml==0.7.1
Django==3.1
django-cors-headers==3.2.1
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
gunicorn==20.1.0
idna==3.3
oauthlib==3.1.1
pycparser==2.21
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2021.3
PyYAML==5.3.1
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
uritemplate==3.0.1
urllib3==1.26.7
whitenoise==5.3.0
